requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1
supabase>=2.3.0
python-dotenv>=1.0.0

# Image analysis
Pillow>=10.3.0
opencv-python-headless>=4.9.0
numpy>=1.26.0

# Audio analysis
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# PDF analysis
PyMuPDF>=1.24.0       # fitz
pdfplumber>=0.11.0

# NLP helpers (optional, for advanced text analysis)
# transformers>=4.40.0
# torch>=2.3.0

# Dev
httpx>=0.27.0
